Revert "Fix - Viewport does not return to highlighted message after returning from thread"
