+++ src/components/InvertedFlatList/BaseInvertedFlatList/index.tsx
@@ -1,8 +1,10 @@
 import type {ForwardedRef} from 'react';
-import React, {forwardRef, useCallback, useImperativeHandle, useRef} from 'react';
-import type {FlatListProps, ListRenderItem, ListRenderItemInfo, FlatList as RNFlatList} from 'react-native';
+import React, {forwardRef, useCallback, useEffect, useImperativeHandle, useMemo, useRef, useState} from 'react';
+import type {FlatListProps, ListRenderItem, ListRenderItemInfo, FlatList as RNFlatList, ScrollViewProps} from 'react-native';
 import FlatList from '@components/FlatList';
-import useFlatListScrollKey from '@hooks/useFlatListScrollKey';
+import usePrevious from '@hooks/usePrevious';
+import getInitialPaginationSize from './getInitialPaginationSize';
+import RenderTaskQueue from './RenderTaskQueue';
 
 // Adapted from https://github.com/facebook/react-native/blob/29a0d7c3b201318a873db0d1b62923f4ce720049/packages/virtualized-lists/Lists/VirtualizeUtils.js#L237
 function defaultKeyExtractor<T>(item: T | {key: string} | {id: string}, index: number): string {
@@ -24,17 +26,56 @@ type BaseInvertedFlatListProps<T> = Omit<FlatListProps<T>, 'data' | 'renderItem'
     initialScrollKey?: string | null;
 };
 
+const AUTOSCROLL_TO_TOP_THRESHOLD = 250;
+
 function BaseInvertedFlatList<T>(props: BaseInvertedFlatListProps<T>, ref: ForwardedRef<RNFlatList>) {
     const {shouldEnableAutoScrollToTopThreshold, initialScrollKey, data, onStartReached, renderItem, keyExtractor = defaultKeyExtractor, ...rest} = props;
-    const {displayedData, maintainVisibleContentPosition, handleStartReached, setCurrentDataId} = useFlatListScrollKey<T>({
-        data,
-        keyExtractor,
-        initialScrollKey,
-        inverted: true,
-        onStartReached,
-        shouldEnableAutoScrollToTopThreshold,
+    // `initialScrollIndex` doesn't work properly with FlatList, this uses an alternative approach to achieve the same effect.
+    // What we do is start rendering the list from `initialScrollKey` and then whenever we reach the start we render more
+    // previous items, until everything is rendered. We also progressively render new data that is added at the start of the
+    // list to make sure `maintainVisibleContentPosition` works as expected.
+    const [currentDataId, setCurrentDataId] = useState(() => {
+        if (initialScrollKey) {
+            return initialScrollKey;
+        }
+        return null;
     });
+    const [isInitialData, setIsInitialData] = useState(true);
+    const currentDataIndex = useMemo(() => (currentDataId === null ? 0 : data.findIndex((item, index) => keyExtractor(item, index) === currentDataId)), [currentDataId, data, keyExtractor]);
+    const displayedData = useMemo(() => {
+        if (currentDataIndex <= 0) {
+            return data;
+        }
+        return data.slice(Math.max(0, currentDataIndex - (isInitialData ? 0 : getInitialPaginationSize)));
+    }, [currentDataIndex, data, isInitialData]);
+
+    const isLoadingData = data.length > displayedData.length;
+    const wasLoadingData = usePrevious(isLoadingData);
     const dataIndexDifference = data.length - displayedData.length;
+
+    // Queue up updates to the displayed data to avoid adding too many at once and cause jumps in the list.
+    const renderQueue = useMemo(() => new RenderTaskQueue(), []);
+    useEffect(() => {
+        return () => {
+            renderQueue.cancel();
+        };
+    }, [renderQueue]);
+
+    renderQueue.setHandler((info) => {
+        if (!isLoadingData) {
+            onStartReached?.(info);
+        }
+        setIsInitialData(false);
+        const firstDisplayedItem = displayedData.at(0);
+        setCurrentDataId(firstDisplayedItem ? keyExtractor(firstDisplayedItem, currentDataIndex) : '');
+    });
+
+    const handleStartReached = useCallback(
+        (info: {distanceFromStart: number}) => {
+            renderQueue.add(info);
+        },
+        [renderQueue],
+    );
 
     const handleRenderItem = useCallback(
         ({item, index, separators}: ListRenderItemInfo<T>) => {
@@ -43,6 +84,19 @@ function BaseInvertedFlatList<T>(props: BaseInvertedFlatListProps<T>, ref: Forwa
         },
         [renderItem, dataIndexDifference],
     );
+
+    const maintainVisibleContentPosition = useMemo(() => {
+        const config: ScrollViewProps['maintainVisibleContentPosition'] = {
+            // This needs to be 1 to avoid using loading views as anchors.
+            minIndexForVisible: data.length ? Math.min(1, data.length - 1) : 0,
+        };
+
+        if (shouldEnableAutoScrollToTopThreshold && !isLoadingData && !wasLoadingData) {
+            config.autoscrollToTopThreshold = AUTOSCROLL_TO_TOP_THRESHOLD;
+        }
+
+        return config;
+    }, [data.length, shouldEnableAutoScrollToTopThreshold, isLoadingData, wasLoadingData]);
 
     const listRef = useRef<RNFlatList | null>(null);
     useImperativeHandle(ref, () => {
@@ -88,5 +142,7 @@ function BaseInvertedFlatList<T>(props: BaseInvertedFlatListProps<T>, ref: Forwa
 BaseInvertedFlatList.displayName = 'BaseInvertedFlatList';
 
 export default forwardRef(BaseInvertedFlatList);
+
+export {AUTOSCROLL_TO_TOP_THRESHOLD};
 
 export type {BaseInvertedFlatListProps};
